// src/api/project_requirement.ts
export type BaseRequirementInfo = {
    title: string;
    content: string;
    tag_id_list: string[];
};

export type RequirementTag = {
    tag_id: string;
    tag_name: string;
    bg_color: string;
};

export type UserRequirementPerm = {
    can_update: boolean;
    can_remove: boolean;
    can_open: boolean;
    can_close: boolean;
}

export type WatchUser = {
    member_user_id: string;
    display_name: string;
    logo_uri: string;
};

export type RequirementInfo = {
    requirement_id: string;
    project_id: string;
    base_info: BaseRequirementInfo;
    issue_link_count: number;
    create_user_id: string;
    create_time: number;
    create_display_name: string;
    create_logo_uri: string;
    update_user_id: string;
    update_time: number;
    update_display_name: string;
    update_logo_uri: string;
    closed: boolean;
    tag_info_list: RequirementTag[];

    kano_excite_value: number;//兴奋型
    kano_expect_value: number;//期望型
    kano_basic_value: number;//基础型
    kano_nodiff_value: number;//无差异
    kano_reverse_value: number;//反向型
    kano_dubiouse_value: number;//可疑数值
    four_q_urgency_value: number;//紧急层度
    four_q_important_value: number;//重要层度

    user_requirement_perm: UserRequirementPerm;

    my_watch: boolean;
    watch_user_list: WatchUser[];
};